requirements.txt
torch
torchaudio
torchvision 
numpy
pandas
matplotlib
scikit-learn
librosa
ffmpeg-python
requests